Add UiStub for test cases
The test cases for adding a new deadline fails as execution of command
is unable to access the Ui during test running.

This essential test cases if left unchanged might cause other issues
to be hidden.

A UiStub that can be referred to by the test cases would resolve the
issue.

Let's
* Add a UiStub with limited functionality for the purpose of testing