Update command execute with assertions
The command class uses an exception to prevent the usage of commands
without parsing the necessary arguments. The command class uses an
exception to re-execution of commands.

The cases represent issues with ordering of function calls, a programmer
error.

This does not follow the convention of using exceptions for user input
errors and assertions for programmer errors.

Let's
* Replace the arguments not parsed exception with an assertion
* Replace the command already executed exception with an assertion