Add comments to justify code duplication
The peer review highlighted enquiries regarding the duplication of
select code in parsing of arguments.

However the code is designed such that it is easily extensible. As such,
the parsing of the arguments are coupled with the command such that
modifications to the command only affect that specific class.

Let's
* Add a comment to justify why the code is structured in such a manner