Add DoWithinPeriod tasks
The set of deadlines, events and todos could be seen as limiting. Users
may find more options of task categories useful. A possible variation of
tasks would be DoWithinPeriod tasks, which indicates tasks that should
be done within a certain period.

Functionally, a DoWithinPeriod task would be identical to an Event task,
as the validation of whether a task is marked as done within a certain
time frame is out of scope. The distinguishing factor would be task
keyword that is displayed.

Notably, the argument format will be identical to an Event task as well.
However the parsing of arguments are to remain separate to avoid
coupling commands together for future changes.

Let's
* Add a DoWithinPeriod task
* Add a AddDoWithinPeriod command
* Add the command to Parser
* Update the version of the application